Delete lost thing image from Firebase Storage if it exists
This commit updates the `_LostThingState` class in the `lost_thing_detail_screen.dart` file. It adds a condition to check if the `imageUrl` property of the `widget.lostThings` object is not empty before attempting to delete the corresponding image from Firebase Storage. This change ensures that the image is only deleted if it exists, preventing unnecessary deletion attempts.

Recent user commits:
- map
- map_inputLocation
- delete lost thing image from firestroge

Recent repository commits:
- feat: Remove web-specific code for setting external user ID
- feat: Enable OneSignal login for non-web platforms
- feat: Add flutter_image_compress dependency for image compression

+++ final_project_front/lib/screens/lost_thing_detail_screen.dart
@@ -98,14 +98,16 @@ class _LostThingState extends State<LostThingDetailScreen>
 
       if (code == 200) {
         // Delete the image from Firebase Storage
-        Uri uri = Uri.parse(widget.lostThings.imageUrl);
-        String fileName = uri.pathSegments[4];
-        FirebaseStorage.instance
-            .ref()
-            .child(fileName)
-            .delete()
-            .then((i) => print("Delete success"))
-            .catchError((error) => print('Failed to delete file: $error'));
+        if (widget.lostThings.imageUrl.isNotEmpty) {
+          Uri uri = Uri.parse(widget.lostThings.imageUrl);
+          String fileName = uri.pathSegments[4];
+          FirebaseStorage.instance
+              .ref()
+              .child(fileName)
+              .delete()
+              .then((i) => print("Delete success"))
+              .catchError((error) => print('Failed to delete file: $error'));
+        }
 
         showAlertDialog('成功', '貼文已刪除', context, success: true, popTwice: true);
       } else if (code == 404) {
